moved jobs to seperate file
